hostapd: Fix channel number in MBSSID RNR after channel switch.
Currently, when channel switch happens in MBSSID case, the
non-transmitting partial profile RNR element of the first few
beacons in the new channel has old channel number.

This is because after channel switch, the new MBSSID and RNR
elements are not updated immediately. During channel switch, the new
MBSSID and RNR elements are not passed to the driver together with
other beacon data. So beaconing first starts with old MBSSID and RNR
data. After channel switch completes and beacon is set again, these
elements are passed to the driver. This delay causes the first few
beacons to contain old RNR and MBSSID elements.

Fix this issue by passing the new MBSSID and RNR elements during
the channel switch itself.

Change-Id: I980d2d056bc96d007c5ec05611327ab315d0938b
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>