hostapd: Fix NULL driv_priv pointer access
In WDS REP, after STA getting associated to root AP, REP AP
interfaces will be up. While a MLD interface is getting
enabled for the first time, bss and driv_priv will be created
in hostapd.

If the MLD interface is failed to enable for some reason,
we need to deinit bss and driv_priv to avoid NULL pointer
access.

Fixes: I2eaffc02782 (hostapd: fix link BSS de-initialize)

Change-Id: I178575437afd43c3025bf27dea2dcf5d6e51b274
Signed-off-by: Monika Korada <quic_koramoni@quicinc.com>