hostapd: Check interface added in bss link de-init
Currently, the links are added to the dl list during driver init and
setup bss when hapd->interface_added is set. However, the below issue is
seen when an item (hapd->link) is not present in the dl list. This
scenario occur during the interface de-init operation where bss link
de-init tries to remove an item which is not present in the list.

The fix is add a check inside bss link de-init to return before the
removal of the mld link from the dl list if hapd->interface_added is
false.

0x00000055560c9610 in dl_list_del.lto_priv.9.lto_priv ()
0x00000055560ca944 in hostapd_bss_link_deinit.lto_priv ()
0x00000055560d3680 in hostapd_bss_deinit ()
0x00000055560d5d90 in hostapd_interface_deinit ()
0x00000055560d5e14 in hostapd_interface_deinit_free ()
0x00000055560c5e40 in hostapd_main ()

Fixes: I7f268b23ff2 ("hostapd: MLO: optimize MLO flows")

Change-Id: Icb2cc84dec02ab17e6a4ec0938dc3a1a68d3f251
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>