hostapd: fix null pointer access during drv->wiphy_list del
The wiphy_list node would be deleted during hostapd interface
disable which would set its list to null. Also, same would be done
as part of nl80211_get_wiphy_data_ap(). This sequence can lead to
double delete of the node.

Currently, to handle this wiphy_list in drv pointer is deleted and
initialized immediately, just to avoid double free in
nl80211_get_wiphy_data_ap which would not be ideal.

To handle gracefully, rather than initializing the wiphy_list
we have to ideally check whether it is null or not before deleting it.
Add changes to handle the same.

Fixes: Ic1339a926 ("hostapd: fix NULL drv pointer access")

Change-Id: Ib73bd754b4760d111c996a69ca5dd9aecd206cab
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>