[WAR] hostapd: fix uninitialized pointer, allowed_freqs access
The pointer, allowed_freqs was uninitialized and
dereferenced from int_array_add_unique API. So it
is set to NULL before the API is called and freed
only if it is valid.

Also, this API is needed only for QCA CLD driver.
ath platform won't support this. Hence protected this
API call with CONFIG_DRIVER_NL80211_QCA flag.

Change-Id: If5e9e2f3bf774a7914920529406778128470af87
Signed-off-by: Monika Korada <quic_koramoni@quicinc.com>