Revert "hostapd: Check interface added in bss link de-init"
This reverts commit 5cbb72e228768fcd67b5878a43cb2640440aa107.

Reason for revert: Currently, the change in this commit works well when hapd->interface_added flag is handled properly. However, it turns out that this flag is not set for mld_ap mode. Currently, adding and removing mld link to the list needs to be handled carefully w.r.t to add link to the driver.

Change-Id: Ibcd671d38bf7cdced8001e8d289ca40de6eb077d