hostapd: Fix mld link dl list crash
Currently, the hapd->link is added to the mld links dl list during
driver init and setup BSS operation. However, the below call trace is
seen when an BSS link item is not present in the dl list and tries to
delete it from the list. This scenario occurs during the interface
deinit operation which calls bss link deinit and tries to remove the
hapd which is not present in the dl list.

The fix is to make sure the link item is added to the dl list after
successful operation of link addition and make sure the mld->num_links
increments only when it is successful. So during bss link deinit
returns if mld->num_links is zero. Since the mld object is shared among
all the links, so the num_links needs to be incremented only when
successful.

Call trace:
        0x00000055560c9610 in dl_list_del.lto_priv.9.lto_priv ()
        0x00000055560ca944 in hostapd_bss_link_deinit.lto_priv ()
        0x00000055560d3680 in hostapd_bss_deinit ()
        0x00000055560d5d90 in hostapd_interface_deinit ()
        0x00000055560d5e14 in hostapd_interface_deinit_free ()
        0x00000055560c5e40 in hostapd_main ()

Change-Id: I8b9a8291aa06ea0ada68c16086717726fdb854dc
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>
(cherry picked from commit c23b4d89d5831a890bb4a08cc2c556da43c62815)