hostapd: Fix channel switch to DFS channel
When we are configuring option channel auto, we are not able to switch to
given dfs channel because when we are trying to move to a dfs channel
interface is disabled and enabled again. When interface is disabled and
enabled we are setting iface's freq and channel to 0 in function
setup_interface2() in case of acs is enabled, now we don't know which
channel we were trying to move. Now ACS will run and interface will be
up in the channel that is suitable.

To fix this issue add flag named is_ch_switch_dfs to check if channel
switch request for dfs channel and we can use this in setup_interface2()
to decide that we have to set iface's freq and channel to 0 or not.
This way iface's freq and channel will retain the values while channel
switch to dfs channel when acs is enable.

Change-Id: I32b1a3453a89b2d2c8adf9fe182c526f53818920
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>