hostapd: Update DTIM information properly in per-STA profile
The DTIM information in the per-STA profile is being incorrectly
updated. The DTIM period is being updated in the LSB byte of the
DTIM information (2 bytes), which is intended for the DTIM count.

Fix by updating the DTIM period and DTIM count information
properly to the MSB and LSB bytes of the DTIM information
respectively.

Change-Id: Iec6a2d02f0a494d815ede8ada1b58c6f42daa6fd
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>