wpa-supplicant: Check MLD address of ML IE in scan result
In a testcase, where is a rootAP operating on 3 links (2, 5 and 6 GHz)
with 15 AP MLDs with a monitor MLD, and WDS repeater operating on
2 links (5 and 6 GHz) with 15 AP MLDs and 1 Station MLD, initially,
the rootAP and repeater connect on a single link on 6 GHz, while the
rootAP undergoes CAC on 5 Ghz link. Once the CAC is complete on rootAP,
the repeater does not deauthenticate and reconnect back on 2 links.

In the wpa_supplicant side on the repeater, in wpas_scan_for_rnr_entries,
the scan results have the BSSID of the 6 GHz link and the BSSID present
in wpa_s->current_bss->bssid is that of the RootAP's MLD address.
As a result, rootAP's RNR IE update done on 6 GHz link is not looked up,
which prevents the repeater from getting deauthenticated from the rootAP.

Fix issue by checking MLD address of ML IE present in the scan result.
This will ensure that the scan result with the RNR IE update on 6 GHz
link, will not get dropped and make the WDS Repeater disconnect from
the RootAP, and reconnect back on 2 links.

Change-Id: I05dcd4996a123e08209e3d6522cf75a03a8d5d6e
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>
Signed-off-by: Pavithra Ganesan <quic_pavigane@quicinc.com>