hostapd: Check MBSSID/RNR fault with 6G CSA.
When channel is switched in 6 GHz in SLO case with hostapd_cli command,
a segmention fault is observed.

In hostapd_build_beacon_data, the MBSSID's beacon offset is
filled without checking if MBSSID/RNR element is added in beacon.
Since MBSSID/RNR is not configured in SLO case, a segmentation
fault is observed.

Fix the issue by checking the MBSSID/RNR element is present in beacon
before filling.

Change-Id: Idfc44b41e667aa84eebefd4d8bb02d9b0b03d768
Signed-off-by: Pavithra Ganesan <quic_pavigane@quicinc.com>