QSDK: hostapd: Add support for configuring interface offload type per netdevice
Add support to configure interface offload type through the configuration
file. This configuration allows to choose datapath offload type on a per
netdevice basis. Configure the datapath offload type using option 'ppe_vp'
in the uci under the wifi iface and then option would be configured to
the driver using a vendor command. The hostapd reads this configured value
from the config file and then sends the vendor command to configure the
driver to use the right datapath offload as instructed by the user

Patch-depdendency: None

Change-Id: Ibf3071402287a2923de1db2bbcd5f3a5b34592aa
Signed-off-by: Pradeep Kumar Chitrapu <quic_pradeepc@quicinc.com>