wpa_supplicant: Fix memleak issue
Currently, when the number of associated links is less than the
requested link, the timer that runs every 5 seconds to fetch the scan
entry does not properly free the scan results memory.
This leads to a memory leak. To address this issue, ensure that the
memory allocated for scan results is properly freed after use.

Fixes: fc72724db0fc ("[WAR] wpa_supplicant: Add changes to reassociate if")

Change-Id: If93d6be8064ae755f63c7ef8f43f6fd56793058c
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>