KW: hostapd: Fix null pointer in FT RRB RX request processing
Currently, hostapd crashes during the processing of an FT
RRB RX request if it encounters the absence of an MLD
address.

This issue arises because the wpa_ft_rrb_rx_request()
function dereferences the mld_addr address for printing
without prior validation for a null value.

To address this issue, the fix involves relocating the
print function call to the 'if' block that executes
only when mld_addr is not null.

Additionally, change the return type of the
hostapd_wpa_auth_config_sanitize() function to void, as
it does not return anything, to avoid compiler warnings.

Patch-work: None
Patch-dependency: r04-007-hostapd-fix-wpa_auth-conf-during-reconfig.patch
r04-009-02-hostapd-add-FT-over-DS-support.patch
Feature-code: MLO

Fixes: I0b18c3f1628 ("hostapd: add FT over DS support")

Change-Id: Ic4de0b368086b626602bed32372f75adbe5ed4cd
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>