hostapd: Fix PTK calculation error in FT over Air roaming
When FT roaming is initiated on a link whose address is not the MLD
address, a PTK mismatch is observed against the PTK calculated in
the STA. This was found to be due to the usage of link addresses of
AP and STA instead of MLD addresses in the PTK derivation.

To fix this issue, call wpa_auth_set_ml_info to set proper MLD addresses
in the sm that will be used for the PTK derivation.

In wpa_ft_validate_reassoc, allocate the assoc_data_link based on the
number of valid links instead of using the n_mld_affiliated_links from
the sm as it would have values from the previous STA.

Also, reset the n_mld_affiliated_links in the sm to 0 when receiving the
(re)assoc request, so that the number of links doesn't get incremented
beyond the actual number of links every time a previously connected STA
sends an FT auth request again.

Fixes: If5d08c2d2 (hostpad: Rework MLO support based on upstream design)

Change-Id: I38701304ba9e850be3a0f027228d4130b3ea52c6
Signed-off-by: Amith A <quic_amitajit@quicinc.com>