KW: hostapd: Fix null pointer in filling critical update capab
Currently, hostapd experiences crashes when a null pointer
is passed to the hostapd_critical_update_capab() function
while filling in the Critical Update Capabilities during
the generation of a probe response.

The issue arises because the function
hostapd_critical_update_capab() attempts to dereference the
'hapd' pointer without first checking if it is null. This
oversight can lead to a null pointer dereference, causing
the function to fail or hostapd to crash unexpectedly.

Address this issue by implementing a null check for the
'hapd' variable before any operations that involve
dereferencing it.

Patch-work: none
Patch-dependency: r04-006-02-hostapd-update-critical-bit-and-bpcc-on-probe-assoc-.patch
Fixes: I8b4ec127367 ("hostapd: update critical bit and bpcc on probe/assoc response frame")

Change-Id: I5411a8489e8c365b77c27714541603f756fa1f0e
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>
(cherry picked from commit 3127fc98c14f02f781c3402b692756abff967756)