KW: hostapd: fix null pointer in sending association response
Currently, hostapd experiences crashes when the
send_assoc_resp() function is called with a null 'sta'
pointer.

The issue arises because the send_assoc_resp() function
attempts to dereference the pointer 'sta' without first
verifying whether it is null.

Resolve this issue by implementing a null check for the
'sta' pointer before accessing it. This additional check
will ensure that the code does not attempt to access or
dereference a null pointer.

Patch-work: None
Patch-dependency: r04-009-hostapd-handle-intel-client-dhcp-failure.patch
Fixes: Iaad7fbe3b3f ("hostapd: Fix association issue for specific clients")

Change-Id: I565e7dee9faca4f6f41e256bb6b9696f816bb8f4
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>