hostapd: Add support to switch channel when CAC is in progress
Prior initiating communication in a DFS channel, there should be a
monitoring(known as Channel Availability Check or CAC) of RADAR in
that channel for a minimum of 600 seconds if it is a weather RADAR
channel and 60 seconds for other DFS channels. During this period,
the AP does no beaconing. Hence no channel switch was allowed during
this period. But if a user wishes to switch to a Non-DFS channel and
start transmission immediately, there is no provision to do that
currently.

In order to enable this provision, extend the existing hostapd_cli
chan_switch command to allow channel switch even when the AP is in
CAC state. To do this, first disable the interface and this should
abort the CAC in progress. Next configure the iface with the new
params and then enable the interface back.

Patch-dependency: None
Patch-work: https://w1.fi/cgit/hostap/commit/?h=pending&id=5cb6747f9721accaf264bd9935ca257183799877

Change-Id: I51eb5c44c6ca31ca3029c05ccb556f1bb2315b41
Signed-off-by: Amith A <quic_amitajit@quicinc.com>