hostapd: SSID protection in 4-way handshake on AP
Add support for SSID protection in 4-way handshake based on the mechanism
added in IEEE 802.11REVme/D6.0. This is a mitigation against
CVE-2023-52424 (a.k.a. the SSID Confusion Attack).
This functionality is disabled by default and can be enabled with
ssid_protection=1.

wpa_supplicant: STA side support to parse ssid protection

Parse SSID protection flag in STA and use it in 4-way handshake.

Change-Id: I43401284b3d96ad1680eb98113541630e0076d28
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>