hostapd: Fix MCS NSS map in beacon for BW less than of equal to 80MHz
When an AP is operating in a bandwidth less than or equal to 80 MHz, it
still advertises support for 320MHz in 6 GHz in the EHT PHY Cap IE. When
a client device identifies this support, it expects the MCS & NSS map to
include maps of 320MHz and 160MHz in the EHT Cap IE. As the AP is
operating in a bandwidth less than or equal to 80MHz, it never includes
the MCS & NSS map of 160MHz. This results in an association failure at
the client.

To fix this issue, enable the 320MHz support in the EHT PHY Cap IE only
if the operating bandwidth is 320MHz.

Change-Id: Id5f982f725c52ba0fd2a623901f9c0d5f47b6632
Signed-off-by: Amith A <quic_amitajit@quicinc.com>
(cherry picked from commit 4b0538efc85d3be67c995adc01a37bcacde799e5)