Update partner linkId for proper partner link reset.
Currently, when a STA is trying to associate with one link
(say 6 GHz) and due to EAPOL failure (M3 failure) at the STA side,
the STA is not responding leading to deauth. The STA then tries to
associate immediately using another link (say 5 GHz).

At present, we still hold the reference to the sta data until the
maximum inactivity timer (currently 5 seconds) expires, as the STA
may try to associate immediately. When we receive an authentication
request on another link before the expiry of the inactivity timer,
we update the STA's linkid and clear the partner link data as part
of ap_free_sta(). However, the reference is not set to NULL due to
the condition (psta->mld_assoc_link_id == lsta->mld_assoc_link_id)
in set_for_each_partner_link_sta(). As a result, when we receive an
auth request, the sta reference is still available but the data is
cleared after the linkid update, leading to a null pointer deference
at wpa_auth_sm_event().

Fix the issue by updating the linkId for all partner links.
This ensures that the other link context is cleared as part of
set_for_each_partner_link_sta().

Change-Id: I98722656fa762270805fcaa2c206cd9489de4f69
Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>