hostapd: Fix memory leak in hostapd_get_reg_max_eirp API
In hostapd_get_reg_max_eirp API, hostapd_channel_data is allocated
dynamically. But it is not freed when an error condition is reached. This
results in a memory leak.

To fix this memory leak, call free_ieee_ordered_chan_list to free the
dynamically allocated hostapd_channel_data.

Fixes: I13e7180a5b9 ("hostapd: Add AFC power compliant TPE")
Patch-dependency: r32-002-hostapd-Add-AFC-power-compliant-TPE-for-Standard-pow.patch

Change-Id: I733933abc30bd1f979e7ffdcc7deaee5bfd0f5a8
Signed-off-by: Amith A <quic_amitajit@quicinc.com>