hostapd: Make dpp configurator parameter as global config
In multi-BSS configuration, if any DPP presence announcement received from
STA, any BSS can respond to it. But as of now DPP configs are interface
specific. Due to that if any interface which does not have DPP configs
responds first, the DPP process will fail, since it don't have the
configs.

Make DPP configurator parameters as global configs so it can be visible
to all the interfaces in hostapd. The bootstrap params are currently
as part for dpp global structure. So the configs also should be part of
dpp global structure.

Patch-dependency: None
Tag: DPP

Change-Id: I197769750baedf5593606b604c6a526ec7b9648f
Signed-off-by: Arunpandi Kannan <quic_arunpand@quicinc.com>