UPSTREAM: hostapd: Extend frequency configuration to handle 6 GHz channel 2
In hostapd_set_freq_params(), if center_segment0 is 2, call
ieee80211_chan_to_freq() with operating class 136 instead of 131.

This is needed because, channel 2 is an exception in the 6 GHz band. It
comes before channel 1 and is part of operating class 136.

Channels order in 6 GHz:
    2 (Operating Class 136)
    1   5   9 ....  (Operating Class 131)

Fixes: None
Patch-dependency: None
Patch-work: None (Already available in upstream)

Change-Id: I110fc341fdefbd0a2fda63fc2481b02db254167c
Signed-off-by: Amith A <quic_amitajit@quicinc.com>