wpa_supplicant: Fix skip scan when noscan is enabled for mesh
Currently mesh is being brought up in channel 36 when
it is configured in channel 40. This is because during HT40
scan when there are no beacons in the configured primary channel
(say channel 40), it will switch the primary and secondary channels
and it results in bringing up the vaps in channel 36.

This leads to issue when both AP+Mesh is brought up in channel 40.
As mesh is coming up in channel 36, AP is not able to come up in
configured channel 40.

Fixing this by skipping the scan part when noscan is enabled during mesh
bring up. This fix which was already present has been modified during the rebase.

Fixes: I3ab1bdfa525e (wpa_supplicant: Rebase mesh changes on OpenWRT hostapd)

Patch-dependency: None
Patch-work: None

Change-Id: I7a4c562e3650badede1029846fb749b92eba264c
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>