hostapd: Fix station disconnect during SA query
The station disconnects and reconnects upon receiving an association
request frame from a connected station without processing the SA query.

When the AP receives an association request from a connected station,
it sends an SA query, starts the SA query timer, and responds with
the status code WLAN_STATUS_ASSOC_REJECTED_TEMPORARILY.

If the status code is not successful, hostapd processes the ML info,
which leads to the deletion of the peer and the addition of a new peer.

Add the change in handle_assoc so that hostapd handles the ML info
when the response code is successful.

Patch-dependency: None
Patch-work: None

Change-Id: Ibe1696415518747ceda1a3d0db2c8b125eeb969f
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>