hostapd: Assign unique sta id for each sta in the mbssid
If mbssid is enabled, the sta_aid space will be shared by all the vifs
in the mbss. With this change, each sta in the mbssid will get a unique
sta aid irrespective of the bss. This can resolve the trigger frame
failure which is observed in fw.

Patch-dependency: none
Patch-work: none

Change-Id: I6b16e1f9f15f19eac4ce8d287a8abe945f78f9a5
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Signed-off-by: Naveen S <quic_naves@quicinc.com>