scan: pass correct link ID in all cases
In hostapd, when a scan was initiated, the link ID parameter was not
populated in all scenarios, such as ACS. Additionally, each caller of
hostapd_driver_scan() provided the link ID. However, since
hostapd_driver_scan() has access to the hapd pointer, it can populate the
link ID itself.

And from wpa_supplicant, link ID was passed as 0 which does not seem to be
correct. Fix that as well.

Patch-dependency: None
Patch-work: [PATCH] scan: pass correct link ID in all cases
Tag: Scan

Change-Id: Ifbc2f237a0fbe57121eebc14e91947edeefba9ec
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>