wpa_supplicant: Initate auth for non EHT Station
While establishing non EHT connections, station loops in scanning
and not able to associate with non EHT AP.

So added validation to check whether the station is EHT/MLD or not.

Incase of non EHT station, supplicant proceeds to authenticate
with the available scan result else attempt to retry scan and
find all configured links in mld so that supplicant will be able
associate on all configured links.

Change-Id: I27ed453ff648e317910b455e640087d8d1ea977b
Signed-off-by: Pavithra Ganesan <quic_pavigane@quicinc.com>