hostapd: Rectify multiple PSD in TPE of 6 GHz SP AP
When a 6 GHz AP operates on Standard Power in 320 MHz, the presence
of multiple PSDs is observed in the 11AX Transmit Power envelope IE,
even though all of the PSDs are the same.

In hostapd_is_freq_present_in_freq_obj, when AFC PSD is extracted from
the AFC frequency objects, the AFC PSD will be in multiples of 100.
This PSD should be divided by 100 when it is filled in the TPE IE.
Since this is not done, if a max_psd in the frequency object is 2300,
it is out of range when stored in a signed 8-bit integer.

To fix this issue, divide the max_psd by 100 before filling it in the
TPE IE.

Tag: 6G_REGULATORY
Fixes: I276e5113e28 ("hostapd: Add AFC power compliant TPE for Standard power AP")

Change-Id: I50f82d517ff81ea1665ba450242f463e038247bd
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>