wpa-supplicant: Check MLD address when comparing rnr_entries in scan entry
In a testcase where a rootAP operates on two links (5 GHz and 6 GHz) MLD,
and a WDS client also operates on two links (5 GHz and 6 GHz) MLD, the
initial connection is established on a single link at 6 GHz while the
rootAP undergoes CAC on the 5 GHz link. Once the CAC is complete on the
rootAP, the WDS client does not deauthenticate and reconnect on both
links.

On the wpa_supplicant side of the WDS client, within the
wpas_scan_for_rnr_entries function, the rootAP's RNR entry is checked
for updates. If any changes are detected, a deauthentication is
triggered. The scan results are iterated, and the scan entry of the rootAP
is fetched for this purpose. This fetching process involves comparing the
rootAP's MLD address in the scan results with the BSSID present in
wpa_s->current_bss->bssid. However, the BSSID in wpa_s->current_bss->bssid
is the link ID of the currently connected link. As a result, the RNR entry
update is never looked up, and the WDS client remains connected without
disconnecting.

To fix this issue, compare wpa_s->bssid, which contains the MLD address of
the BSS, against the MLD address in the scan results. This ensures that
the scan result with the RNR IE update on the 6 GHz link is not dropped,
causing the WDS client to disconnect from the rootAP and reconnect on both
links.

Fixes: I05dcd4996a1 ("wpa-supplicant: Check MLD address of ML IE in scan
		     result")
Signed-off-by: Amith A <quic_amitajit@quicinc.com>

Change-Id: Ic0eb6d602543cd05385acf5e2b06cc0e7cf40e65