WAR: QSDK: hostapd: Fix wpa_supplicant sigterm crash
With Mesh configuration, wpa_supplicant crashes after multiple
wifi up/down.

In eloop_handle_signal, the callback handler is getting invoked and
accesses the eloop.signals which is already freed in previous instance. Since eloop is global pointer and signals memory is not set to NULL, it leads to invalid memory access.

Fix this by setting the signals pointer to NULL after it is freed in
eloop_destroy callback. Once Ucode change is merged, this gerrit can be reverted.

Change-Id: I2fea0c8aa45110c002e8aee21149e64441f56513
Signed-off-by: Karthik <quic_karm@quicinc.com>