hostapd: Fix Number of simultaneous links in MLE during CAC
Currently, the maximum number of simultaneous links in MLE is
advertised as `num_links - 1`, where `num_links` is the number
of links added to the MLD AP. However, when the 5 GHz band is
in CAC timeout, this results in an incorrect value being
advertised for the maximum number of simultaneous links in
MLE, as the 5 GHz link is not active.

For example, an AP MLD with 3 links (2 GHz, 5 GHz (in CAC timeout)
and 6 GHz) during bringup has `num_links` set to 3.
Consequently, the maximum number of simultaneous links
in MLE is advertised as 2 according to the current code,
despite the 5 GHz link being in CAC timeout.

To Fix this issue, introduce an API to retrieve the active links
of the MLD AP and use it to set the value for the maximum
number of simultaneous links in MLE.

Patch-dependency: None

Patch-work: [PATCH] hostapd: Fix Number of simultaneous links in MLE during CAC

Change-Id: I444d2607b1256862f1c67fbc82a7d2ddc2ca1af5
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>