QSDK: hostapd: Fix center frequency set after CAC completion
When CAC is completed, the starting frequency is set to a value
calculated based on the channel and width in which CAC is completed.
This is done assuming that the AP would be moving to the channel in
which CAC was completed. However, in case of background CAC, no channel
switch would be done and hence changing the starting frequency will
lead to an incorrect configuration. When radar is hit somewhere in the
range of the actual start frequency after such an update, it is ignored
as it doesn't fall under the current operating frequency range as per
the updated configuration.

In order to fix this issue, set the starting frequency in the conf
based the CAC completed frequency only if the CAC completion status is a
success. For background CAC, the success status is set to 0 and hence
the starting frequency is never updated.

Fixes: I2c2ab4dcea0 ("QSDK: hostapd: Fix center frequency after channel switch")

Change-Id: Ib108df8a6529923eb072e56041d7fa57ea9367e7
Signed-off-by: Amith A <quic_amitajit@quicinc.com>