hostapd: Update correct puncturing bitmap in radar event
When radar event is received, the radar bitmap is not correctly
updated and puncturing is not applied.

This is because, the radar bitmap is not considered when setting
puncturing bitmap and during channel switch notify events,
puncturing bitmap is not received correctly.

Fix this issue by updating the puncturing bitmap with radar bitmap
when radar event is received and also receive the correct puncturing
bitmap from driver during channel switch notify events.

Patch-depdendency: q02-055-hostapd-ru-puncturing-for-dfs-channels.patchr 30-001-hostapd-add-puncturing-bitmap-to-frequency-parameter.patch

Patch-work:

Feature: PUNCTURING

Change-Id: Ic9340626612c0edec94df5f1618df481fd207082
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>