UPSTREAM:HOSTAPD: Add QCA_WLAN_VENDOR_ATTR_AFC_RESP_HW_IDX to afc response enum
In a scenario where a wiphy contains more than one 6 GHz radio,
the AFC response cannot be routed to the respective radio with the
netdev in a single wiphy architecture.

Therefore, add a new identifier ‘QCA_WLAN_VENDOR_ATTR_AFC_RESP_HW_IDX’
in the qca_wlan_vendor_attr_afc_response, which can be used to route
the AFC response to the respective radio for writing to shared memory.

Tag: 6G_REGULATORY
Patch-dependency: r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch

Change-Id: I69b51825c68094936679bb60ea1f4f2a02984903
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>